Change the date format in agent/base to satisfy ES needs
ES wants date formats like <yyyy>-<mm>-<dd>T<HH>:<MM>:<SS>
with a T separating the date from the time.

Pbench has been using _ instead of T and the indexing script has
been doing the conversion. But there is no reason to stick to the
old format, so change all the dates in agent/base to use T instead
of _.

Fix the unit tests (that's the biggest change).