Index uperf result data properly
WARNING: This is not finished - do not merge!

This commit concerns itself only with the lowest level result.json files (those in
the <iteration>/sample-NN/ directories.

There are problems with timestamps: the timestamp in each document needs to be
converted (I think). Also, the index needs to use the experiment timestamp from
the run doc, not the timestamp from an individual data point.

The unit tests are failing (starting with 7.6). The changes have been tested only in
a small harness outside the main index-pbench script.